[Core] improve overflow safety of AdaptiveFee (16, 150)

+++ src/core/contracts/interfaces/IAlgebraFactory.sol
@@ -32,14 +32,14 @@ interface IAlgebraFactory {
   event FarmingAddress(address indexed newFarmingAddress);
 
   event FeeConfiguration(
-    uint32 alpha1,
-    uint32 alpha2,
+    uint16 alpha1,
+    uint16 alpha2,
     uint32 beta1,
     uint32 beta2,
     uint16 gamma1,
     uint16 gamma2,
     uint32 volumeBeta,
-    uint32 volumeGamma,
+    uint16 volumeGamma,
     uint16 baseFee
   );
 
@@ -105,14 +105,14 @@ interface IAlgebraFactory {
   function setVaultAddress(address _vaultAddress) external;
 
   function setBaseFeeConfiguration(
-    uint32 alpha1,
-    uint32 alpha2,
+    uint16 alpha1,
+    uint16 alpha2,
     uint32 beta1,
     uint32 beta2,
     uint16 gamma1,
     uint16 gamma2,
     uint32 volumeBeta,
-    uint32 volumeGamma,
+    uint16 volumeGamma,
     uint16 baseFee
   ) external;
 }
